requirements.txt
alembic==1.8.1
autopep8==1.7.0
bcrypt==4.0.0
black==22.12.0
blinker==1.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.5
contourpy==1.1.0
cycler==0.11.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.1
fonttools==4.42.1
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.7.3
mypy-extensions==0.4.3
numpy==1.25.2
packaging==23.1
pathspec==0.10.3
Pillow==10.0.0
platformdirs==2.6.0
praw==7.7.1
prawcore==2.4.0
pycodestyle==2.9.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.6
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.40
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
update-checker==0.18.0
urllib3==2.1.0
websocket-client==1.6.4
Werkzeug==2.3.0
WTForms==3.0.1
zipp==3.8.1
